requirements.txt
torch==2.4.1
torchvision==0.19.1
torchmetrics==0.11.4
accelerate==0.20.3
Pillow==9.4.0
tensorboard==2.13.0
timm==1.0.9
scikit-image==0.21.0
diffusers==0.18.2
huggingface-hub==0.24.7
einops==0.6.1
mmcv==1.7.0
came-pytorch==0.1.3
yapf==0.40.1
editdistance
pytesseract